Fix for ST3: hardcode settings name
In ST3, __name__ returns Inc-Dec-Value.inc_dec_value instead
of just inc_dec_value. I'm not sure why this is, but it prevents
us from easily using __name__ to get the name of the settings.
I chose hardcoding instead of trying to do an if(version) kind of
thing; not sure if this is the best aproach.